Function wasmer_runtime_c_api::instance::wasmer_instance_context_data_get

#[no_mangle]
pub extern "C" fn wasmer_instance_context_data_get(
    ctx: *const wasmer_instance_context_t
) -> *mut c_void

Gets the data that can be hold by an instance.

This function is complementary of wasmer_instance_context_data_set(). Please read its documentation. You can also read the documentation of wasmer_instance_context_t to get other examples.

This function returns nothing if ctx is a null pointer.